1.0.0

+++ events/riffyEvents/track/queueEnd.js
@@ -3,11 +3,7 @@ const client = require("../../../Aria");
 client.riffy.on("queueEnd", async (player) => {
     const channel = client.channels.cache.get(player.textChannel);
     
-    if (player.message) try {
-        await player.message.delete().catch((e) => {});
-    } catch (err) {
-        logger(err, "error");
-    }
+    if (player.message) await player.message.delete().catch((e) => {});
 
     if (player.isAutoplay === true) {
         player.autoplay(player)
@@ -15,4 +11,4 @@ client.riffy.on("queueEnd", async (player) => {
         player.destroy();
         channel.send("\`⌛️\` | Queue has ended.");
     }
-})
+});
