+++ events/riffyEvents/track/trackEnd.js
@@ -4,9 +4,5 @@ const client = require("../../../Aria");
 client.riffy.on("trackEnd", async (player) => {
     if (!player) return;
 
-    if (player.message) try {
-        await player.message.delete().catch((e) => {});
-    } catch (err) {
-        logger(err, "error");
-    }
-})
+    if (player.message) await player.message.delete().catch((e) => {});
+});
